pgmaster-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt>=5.3.0
stripe
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==23.4
pillow==10.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0
django-timezone-field==6.1.0
pytz==2023.3
flake8==6.1.0
black==23.12.0
django-celery-beat==2.5.0
reportlab==4.0.9
openpyxl>=3.1.0
google-auth
google-auth-oauthlib

